Moving `ShelfBrowse` into `lib/search/presenters/record/catalog`.
