+++ lib/search/presenters.rb
@@ -5,7 +5,6 @@ end
 require "search/presenters/affiliations"
 require "search/presenters/icons"
 require "search/presenters/search_options"
-require "search/presenters/shelf_browse"
 require "search/presenters/record"
 
 module Search::Presenters
